+++ graphein/protein/graphs.py
@@ -395,8 +395,10 @@ def select_chains(
 def initialise_graph_with_metadata(
     protein_df: pd.DataFrame,
     raw_pdb_df: pd.DataFrame,
-    pdb_id: str,
     granularity: str,
+    name: Optional[str] = None,
+    pdb_code: Optional[str] = None,
+    pdb_path: Optional[str] = None,
 ) -> nx.Graph:
     """
     Initializes the nx Graph object with initial metadata.
@@ -405,17 +407,35 @@ def initialise_graph_with_metadata(
     :type protein_df: pd.DataFrame
     :param raw_pdb_df: Unprocessed dataframe of protein structure for comparison and traceability downstream.
     :type raw_pdb_df: pd.DataFrame
-    :param pdb_id: PDB Accession code.
-    :type pdb_id: str
     :param granularity: Granularity of the graph (eg ``"atom"``, ``"CA"``, ``"CB"`` etc or ``"centroid"``).
         See: :const:`~graphein.protein.config.GRAPH_ATOMS` and :const:`~graphein.protein.config.GRANULARITY_OPTS`.
     :type granularity: str
+    :param name: specified given name for the graph. If None, the PDB code or the file name will be used to name the graph.
+    :type name: Optional[str], defaults to ``None``
+    :param pdb_code: PDB ID / Accession code, if the PDB is available on the PDB database.
+    :type pdb_code: Optional[str], defaults to ``None``
+    :param pdb_path: path to local PDB file, if constructing a graph from a local file.
+    :type pdb_path: Optional[str], defaults to ``None``
     :return: Returns initial protein structure graph with metadata.
     :rtype: nx.Graph
     """
+
+    assert (pdb_code and not pdb_path) or (not pdb_code and pdb_path), (
+        "Either a PDB ID or a path to a local PDB file"
+        " must be specified to read a PDB"
+    )
+
+    # Get name for graph if no name was provided
+    if not name:
+        if pdb_path:
+            name = get_protein_name_from_filename(pdb_path)
+        else:
+            name = pdb_code
+
     G = nx.Graph(
-        name=pdb_id,
-        pdb_id=pdb_id,
+        name=name,
+        pdb_code=pdb_code,
+        pdb_path=pdb_path,
         chain_ids=list(protein_df["chain_id"].unique()),
         pdb_df=protein_df,
         raw_pdb_df=raw_pdb_df,
@@ -553,6 +573,7 @@ def compute_edges(
 
 def construct_graph(
     config: Optional[ProteinGraphConfig] = None,
+    name: Optional[str] = None,
     pdb_path: Optional[str] = None,
     uniprot_id: Optional[str] = None,
     pdb_code: Optional[str] = None,
@@ -574,10 +595,12 @@ def construct_graph(
 
     :param config: :class:`~graphein.protein.config.ProteinGraphConfig` object. If None, defaults to config in ``graphein.protein.config``.
     :type config: graphein.protein.config.ProteinGraphConfig, optional
-    :param pdb_path: Path to ``pdb_file`` to build graph from. Default is ``None``.
-    :type pdb_path: str, optional
-    :param pdb_code: 4-character PDB accession pdb_code to build graph from. Default is ``None``.
-    :type pdb_code: str, optional
+    :param name: an optional given name for the graph. the PDB ID or PDB file name will be used if not specified.
+    :type name: str, optional
+    :param pdb_path: Path to ``pdb_file`` when constructing a graph from a local pdb file. Default is ``None``.
+    :type pdb_path: Optional[str], defaults to ``None``
+    :param pdb_code: A 4-character PDB ID / accession to be used to construct the graph, if available. Default is ``None``.
+    :type pdb_code: Optional[str], defaults to ``None``
     :param uniprot_id: UniProt accession ID to build graph from AlphaFold2DB. Default is ``None``.
     :type uniprot_id: str, optional
     :param chain_selection: String of polypeptide chains to include in graph. E.g ``"ABDF"`` or ``"all"``. Default is ``"all"``.
@@ -595,8 +618,13 @@ def construct_graph(
     :param graph_annotation_funcs: List of graph annotation function. Default is ``None``.
     :type graph_annotation_funcs: List[Callable]
     :return: Protein Structure Graph
-    :type: nx.Graph
+    :rtype: nx.Graph
     """
+
+    assert (pdb_code and not pdb_path) or (not pdb_code and pdb_path), (
+        "Either a PDB ID or a path to a local PDB file"
+        " must be specified to construct a graph"
+    )
 
     # If no config is provided, use default
     if config is None:
@@ -648,6 +676,7 @@ def construct_graph(
             raw_df,
             chain_selection=chain_selection,
             granularity=config.granularity,
+            insertions=config.insertions
         )
         progress.advance(task2)
 
@@ -656,7 +685,9 @@ def construct_graph(
         g = initialise_graph_with_metadata(
             protein_df=protein_df,
             raw_pdb_df=raw_df.df["ATOM"],
-            pdb_id=pdb_code if pdb_code is not None else uniprot_id,
+            name=name,
+            pdb_code=pdb_code,
+            pdb_path=pdb_path,
             granularity=config.granularity,
         )
         # Add nodes to graph
@@ -677,6 +708,7 @@ def construct_graph(
             get_contacts_config=None,
         )
         progress.advance(task4)
+        
     # Annotate additional graph metadata
     if config.graph_metadata_functions is not None:
         g = annotate_graph_metadata(g, config.graph_metadata_functions)